requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
db-dtypes==1.5.0
fastapi==0.128.6
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.16.0
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-gbq==0.33.0
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.54.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xgboost==3.1.3
